+++ test/scratch/consumer_fetchcontent/main.c
@@ -13,9 +13,9 @@ main(void)
 
     if (p99_histogram_value_at_p50(&histogram, &value))
     {
-        printf("p50: %llu ns (%zu events)\n",
+        printf("p50: %llu ns (%llu events)\n",
                (unsigned long long)value,
-               p99_histogram_event_count(&histogram));
+               (unsigned long long)p99_histogram_event_count(&histogram));
         return 0;
     }
